Update README.md
With the post COVID 19 economic restart, flights delays gradually increasing and the result is the financial loss and customer dissatisfaction for airline companies. to resolve this situation and help airlines companies to be more efficient in the flight operation, machine learning models will be implemented to help  airline for fly delay prediction. 
Airlines dataset used is from kaggle and has 539383 instances and 8 different features. The task is to predict whether a given flight will be delayed, given the information of scheduled departure.